requirements.txt
fastapi
uvicorn
python-docx
pydantic
python-multipart
